Added holes function and changed reward system
Made the board smaller to see if the training would go faster. (Not sure yet)

The reward system is now based on 'Tetris AI – The (Near) Perfect Bot's factors and functions.

We now have the reward depending on:
* Bumpiness
* Height
* Holes
* Cleared lines

Cleaned up the code to minimize the warnings.

Also added a get_lines function to send number of lines to print in run_model